refactor: :recycle: move components to folder
