+++ src/lib/components/character-picker.svelte
